Stream S + Stream T: 24 P2 hygiene + 10 tests/build/tooling items
audit-3 P2 and tests/build.

Stream S (P2 hygiene):

S.1 — function_attention.cpp: composed_attention_backward_variable
defensive negative-dim normalise.  Rank check throws if Q/K/V are
<2D after reshape; every -1 / -2 dim literal replaced with explicit
ndim - {1,2}.

S.2 — function_shape.cpp: SliceBackward CPU Int64 index now cached
per (shape, dim, start, end, step, device-type, device-index).
SliceIndexKey/Hash + mutex-guarded unordered_map<Key, Tensor> +
get_or_build_slice_index() helper.  Eliminates per-backward stream
serialisation on GPU dispatch.

S.3 — engine.cpp: anomaly NaN/Inf counter now sums Int64 (was
Float32 — saturated above 2^24 elements, so user-visible "NaN
count: N" was silently wrong for >16M-element tensors).

S.4 — vulkan_backend.{hpp,cpp} + vulkan_ops_misc.cpp: new
vulkan::ensure_fp16_supported() helper called at every FP16 entry
point in ops_misc (Full / Ones / Rand / Randn / MaskedSelect /
Interpolate / ROIAlignForward / ROIAlignBackward / ArgSort / Cast).
Mirrors the R.13 ensure_fp64_supported pattern.

S.5 — vulkan_backend.cpp:1131-1149 + vulkan_caching_allocator.cpp:
caching-allocator now throws typed std::out_of_range for "device
not found" / "ptr not tracked"; outer catch differentiates
out_of_range (try next device) from std::exception (log+rethrow).

S.6 — cuda_kernel_registry.cpp:2986-2993: Conv3d wrapper comment
rewritten to document actual behaviour (kernel throws on
asymmetric, wrapper does not assert).  No comment-vs-code drift.

S.7 — cuda_kernel_registry.cpp: documentation comments at
Dropout / FractionalMaxPool3d / MaxUnpool sections clarifying
which already accept per-axis OutputSize and which are
intrinsically scalar (Dropout p is a probability, not a per-axis
attribute).  ConvTranspose3d non-cuDNN path already uses the
t3d_iso_or_throw helper.

S.8 — cudnn_rnn.cu:217-241: pack_weights lifetime audited; bias
W_ih/W_hh/b_ih/b_hh are const std::vector<Tensor>&, caller-owned.
All async D2D copies share the cudnnRNNForward stream.  No code
change beyond a documenting comment block.

S.9 — vulkan_ops_misc.cpp two-step F64 Cast: submitBatchIfNeeded(…,
true) calls now bracket the inner dispatchCast invocations and the
intermediate is bound to a local so it outlives the in-flight
dispatch.

S.10 — src/backends/rocm/kernels/bfloat16_helpers.hpp (new):
tenzor::rocm::f32_to_bf16_rne(float) — union-pun, __device__
__host__, NaN-preserving, RNE per PyTorch.  Applied to
indexing.hip.cpp atomicAddHelper (with R.11), embedding_bag mean/
max, two masked_fill BF16 sites; math.hip.cpp div / broadcast_div
/ DEFINE_ROCM_SPECIAL_UNARY macro + 20 other ctor sites;
transform.hip.cpp FP8 → BF16 dequant; rocm_nan_helpers.hip.h
safe_f2bf — all switched from the value-truncating hip_bfloat16
ctor to the RNE helper.

S.11 — MPS Conv3d / MaxPool3d / AvgPool3d dilation: kernel
signatures extended with dd/dh/dw; mps_kernel_registry.mm reads
DilationD/H/W with Dilation scalar fallback; output-shape formulas
now dilation-aware ((in + 2*pad - dil*(k-1) - 1)/stride + 1).
mps_extended_ops.mm params array carries dd/dh/dw.  pool3d.metal
shaders gained constant uint& d{d,h,w}_dil parameters and
int(k_idx * d_dil) index math.

S.12 — asgd.cpp / rprop.cpp: load_state_dict gained num_params
guard mirroring the audit-2 P.2 pattern for Adam/AdamW/etc.

S.13 — swa.cpp: averaged_params_ promoted to Float32 master copy
when params are F16/BF16.  apply_to casts back; arithmetic in
state dtype.  Avoids F16 max overflow at n >= 512.

S.14 — optimizer.{hpp,cpp} + pruning.cpp: base Optimizer gained a
step_count() getter backed by atomic<uint64_t> bumped after
step_impl().  Pruning hook idempotence now keys on
optimizer.step_count() (was hook-local atomic — handed distinct
iter numbers to concurrent step()s under threaded DataParallel).

S.15 — normalization.cpp LayerNorm CPU backward: sum_grad_out /
sum_grad_out_normalized promoted to double matching the variance
pass.

S.16 — fsdp2.cpp: Partial{Sum} → Partial{Mean} placement;
manual divide-by-shard_world_size_ dropped.

S.17 — sam.cpp:189-196: strips "sam_rho" key from state before
forwarding to base optimiser's load_state_dict.

S.18 — normalization.cpp LayerNorm backward: branches on
elementwise_affine_; returns {grad_input} only when false, skipping
the placeholder ones-weight multiply.

S.19 — bindings_distributed.cpp: all_gather now takes py::list,
mutates in place, returns None (matches gather / PyTorch).

S.20 — bindings_lite.cpp: LiteRuntime.__call__ and .forward gained
py::call_guard<gil_scoped_release>().

S.21 — bindings_autograd.cpp: func.grad propagates input's
requires_grad flag instead of overriding to true.

S.23 — bindings_nn.cpp: ParameterList / ParameterDict __getitem__
returns shared_ptr<Variable> directly; pybind exposes the Variable
wrapping the same shared storage so mutations propagate.

S.24 — bindings_core.cpp: Tensor.to(...) gained a kwargs overload
(Optional[Device], Optional[DType], copy=false) that routes to
the appropriate C++ overload and forces .clone() when copy=True.

S.22 already covered by R.22 (outer GIL release on gradcheck /
gradgradcheck).

Stream T (build/tooling, tests):

T.2 — tools/scan_audit_markers.py: EXCLUDED_FILENAMES set with
the tool's own basename so the regex source no longer self-flags.

T.3 — scripts/audit_tolerances.py + tools/audit_tolerances_
baseline.txt: --baseline / --update-baseline modes mirroring
lint_test_contract.py.  Baseline lists 330 pre-existing
unjustified tolerances; --strict --baseline exits 0 when all
violations are grandfathered, 1 on new violations.

T.4 — CMakeLists.txt / tests/CMakeLists.txt: 4 new ctest entries
under the "lint" label — AuditMarkersScan, PyiDrift,
ParamGroupSerialization, TensorboardWireFormat.  ToleranceAudit
upgraded to --strict --baseline.

T.5 — tests/CMakeLists.txt: BACKENDS vulkan added to
test_vulkan_layer_norm_saved_stats and test_vulkan_complete_ops
registrations.

T.6 — tests/CMakeLists.txt: BACKENDS multi added to the 6 root
multidtype tests (cholesky_solve, dct, deformable_conv2d,
einsum_dispatch, lobpcg, mfcc).

T.7 — std::srand(42) seeded in MinimalTraining and
ZeRODistributedTestBase SetUp() so std::rand()-based target
generation is reproducible.

T.8 — test_losses_missing_multidtype.cpp: 27 EXPECT_LE(numel, 1)
→ EXPECT_EQ(numel, 1) so a backend returning numel=0 doesn't
pass the assertion.

T.9 — DISABLED_ perf tests: EXPECT_LT(duration, 10000) regression-
guard added with explicit reason comment to all 5 sites
(test_roberta, test_electra, test_async_ops sync+async,
test_vulkan_complete_ops Conv2d+Matmul); wired into the
tenzor_benchmarks umbrella.

T.10 — tolerance justification comments at every site listed in
the plan (test_creation_ops, test_dtype_conversion, test_gradcheck,
test_gradcheck_direct, test_func, test_statistics_multidtype).
audit_tolerances_baseline.txt extended with the new entries under
a # T.10 section.

T.11 — test_onnx_import.cpp uses
temp_directory_path() / ("tenzor_onnx_" + getpid() + "_" +
test_info->name()) so parallel ctest workers don't race on the
shared /tmp/tenzor_onnx_import_test fixture path.

Build: full ninja clean in single -j1 shell after touching
ops.hpp / function.hpp to force dependency refresh on R.1's
double-precision sigs (audit-3 Stream R commits had to settle
their headers before this commit).

+++ src/autograd/engine.cpp
@@ -73,12 +73,15 @@ static void check_for_anomaly(const std::vector<Tensor>& grads,
         auto nan_mask = isnan(grad);
         auto inf_mask = isinf(grad);
 
-        // Sum bool masks as float to check if any anomalies exist
-        auto nan_count = sum(nan_mask.to(DType::Float32));
-        auto inf_count = sum(inf_mask.to(DType::Float32));
+        // S.3 — sum mask in Int64 so the count is exact for tensors above
+        // 2^24 elements. Float32 sum silently saturates beyond that, making
+        // the reported "NaN count: N" diagnostic wrong for 16M+ element
+        // grads. Int64 is also the natural dtype for an integer count.
+        auto nan_count = sum(nan_mask.to(DType::Int64));
+        auto inf_count = sum(inf_mask.to(DType::Int64));
 
-        bool has_nan = nan_count.item<float>() > 0.0f;
-        bool has_inf = inf_count.item<float>() > 0.0f;
+        bool has_nan = nan_count.item<int64_t>() > 0;
+        bool has_inf = inf_count.item<int64_t>() > 0;
 
         if (has_nan || has_inf) {
             // Get demangled function name for readability
@@ -107,14 +110,14 @@ static void check_for_anomaly(const std::vector<Tensor>& grads,
             }
             detail += "]";
 
-            // Anomaly counts
-            detail += "\n  NaN count: " + std::to_string(static_cast<int64_t>(nan_count.item<float>()));
-            detail += "\n  Inf count: " + std::to_string(static_cast<int64_t>(inf_count.item<float>()));
+            // Anomaly counts (S.3 — Int64 to avoid Float32 saturation beyond 2^24).
+            detail += "\n  NaN count: " + std::to_string(nan_count.item<int64_t>());
+            detail += "\n  Inf count: " + std::to_string(inf_count.item<int64_t>());
 
-            // Finite value statistics
+            // Finite value statistics — count in Int64 for the same reason.
             auto finite_mask = logical_not(logical_or(nan_mask, inf_mask));
-            Tensor finite_count_t = sum(finite_mask.to(DType::Float32));
-            float finite_count_val = finite_count_t.item<float>();
+            Tensor finite_count_t = sum(finite_mask.to(DType::Int64));
+            int64_t finite_count_val = finite_count_t.item<int64_t>();
             if (finite_count_val > 0) {
                 auto grad_shape_vec = std::vector<int64_t>(grad.shape().begin(), grad.shape().end());
                 Tensor grad_f32 = (grad.dtype() != DType::Float32) ? grad.to(DType::Float32) : grad;
